basic setup and reorganization
PlayerType added.
PlayerColor and PlayerType both moved to a more appropriate package.
