+++ src/users/Player.java
@@ -1,7 +1,6 @@
 package users;
 
 import components.Card;
-import components.PlayerColor;
 import components.playerCollections.DiscardPile;
 import components.playerCollections.Hand;
 import components.playerCollections.StockPile;
