+++ src/users/PlayerType.java
@@ -1,0 +1,11 @@
+package users;
+
+/**
+ * Used for controlling which Players are AI and which are human.
+ * 
+ * @author Anna Heebsh
+ */
+public enum PlayerType {
+	HUMAN,
+	AI
+}
